+++ libraries/matchers.rb
@@ -1,11 +1,12 @@
 if defined?(ChefSpec)
   chefspec_version = Gem.loaded_specs['chefspec'].version
-  if chefspec_version < Gem::Version.new('4.1.0')
-    define_method = ChefSpec::Runner.method(:define_runner_method)
-  else
-    define_method = ChefSpec.method(:define_matcher)
-  end
+  define_method = if chefspec_version < Gem::Version.new('4.1.0')
+                    ChefSpec::Runner.method(:define_runner_method)
+                  else
+                    ChefSpec.method(:define_matcher)
+                  end
 
+  define_method.call :windows_certificate
   define_method.call :windows_package
   define_method.call :windows_feature
   define_method.call :windows_task
@@ -18,6 +19,83 @@ if defined?(ChefSpec)
   define_method.call :windows_printer
   define_method.call :windows_printer_port
   define_method.call :windows_reboot
+
+  #
+  # Assert that a +windows_certificate+ resource exists in the Chef run with the
+  # action +:create+. Given a Chef Recipe that creates 'c:\test\mycert.pfx' as a
+  # +windows_certificate+:
+  #
+  #     windows_certificate 'c:\test\mycert.pfx' do
+  #       action :create
+  #     end
+  #
+  # The Examples section demonstrates the different ways to test a
+  # +windows_certificate+ resource with ChefSpec.
+  #
+  # @example Assert that a +windows_certificate+ was created
+  #   expect(chef_run).to create_windows_certificate('c:\test\mycert.pfx')
+  #
+  #
+  # @param [String, Regex] resource_name
+  #   the name of the resource to match
+  #
+  # @return [ChefSpec::Matchers::ResourceMatcher]
+  #
+  def create_windows_certificate(resource_name)
+    ChefSpec::Matchers::ResourceMatcher.new(:windows_certificate, :create, resource_name)
+  end
+
+  #
+  # Assert that a +windows_certificate+ resource exists in the Chef run with the
+  # action +:delete+. Given a Chef Recipe that deletes "me.acme.com" as a
+  # +windows_certificate+:
+  #
+  #     windows_certificate 'me.acme.com' do
+  #       action :delete
+  #     end
+  #
+  # The Examples section demonstrates the different ways to test a
+  # +windows_certificate+ resource with ChefSpec.
+  #
+  # @example Assert that a +windows_certificate+ was _not_ deleted
+  #   expect(chef_run).to_not delete_windows_certificate('me.acme.com')
+  #
+  #
+  # @param [String, Regex] resource_name
+  #   the name of the resource to match
+  #
+  # @return [ChefSpec::Matchers::ResourceMatcher]
+  #
+  def delete_windows_certificate(resource_name)
+    ChefSpec::Matchers::ResourceMatcher.new(:windows_certificate, :delete, resource_name)
+  end
+
+  #
+  # Assert that a +windows_certificate+ resource exists in the Chef run with the
+  # action +:acl_add+. Given a Chef Recipe that adds a private key acl to "me.acme.com" as a
+  # +windows_certificate+:
+  #
+  #     windows_certificate 'me.acme.com' do
+  #       private_key_acl ['acme\fred', 'pc\jane']
+  #       action :acl_add
+  #     end
+  #
+  # The Examples section demonstrates the different ways to test a
+  # +windows_certificate+ resource with ChefSpec.
+  #
+  # @example Assert that a +windows_certificate+ was _not_ removed
+  #   expect(chef_run).to add_acl_to_windows_certificate('me.acme.com').with(private_key_acl: ['acme\fred', 'pc\jane'])
+  #
+  #
+  # @param [String, Regex] resource_name
+  #   the name of the resource to match
+  #
+  # @return [ChefSpec::Matchers::ResourceMatcher]
+  #
+  def add_acl_to_windows_certificate(resource_name)
+    ChefSpec::Matchers::ResourceMatcher.new(:windows_certificate, :acl_add, resource_name)
+  end
+
   #
   # Assert that a +windows_package+ resource exists in the Chef run with the
   # action +:install+. Given a Chef Recipe that installs "Node.js" as a
@@ -177,6 +255,31 @@ if defined?(ChefSpec)
   #
   def create_windows_task(resource_name)
     ChefSpec::Matchers::ResourceMatcher.new(:windows_task, :create, resource_name)
+  end
+
+  #
+  # Assert that a +windows_task+ resource exists in the Chef run with the
+  # action +:disable+. Given a Chef Recipe that creates "mytask" as a
+  # +windows_task+:
+  #
+  #     windows_task 'mytask' do
+  #       action :disable
+  #     end
+  #
+  # The Examples section demonstrates the different ways to test a
+  # +windows_task+ resource with ChefSpec.
+  #
+  # @example Assert that a +windows_task+ was disabled
+  #   expect(chef_run).to disable_windows_task('mytask')
+  #
+  #
+  # @param [String, Regex] resource_name
+  #   the name of the resource to match
+  #
+  # @return [ChefSpec::Matchers::ResourceMatcher]
+  #
+  def disable_windows_task(resource_name)
+    ChefSpec::Matchers::ResourceMatcher.new(:windows_task, :disable, resource_name)
   end
 
   #