+++ providers/task.rb
@@ -17,14 +17,17 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 #
+use_inline_resources if defined?(use_inline_resources)
 
 require 'chef/mixin/shell_out'
+require 'rexml/document'
+
 include Chef::Mixin::ShellOut
 
 use_inline_resources
 
 action :create do
-  if @current_resource.exists && (!(task_need_update? || @new_resource.force))
+  if @current_resource.exists && !(task_need_update? || @new_resource.force)
     Chef::Log.info "#{@new_resource} task already exists - nothing to do"
   else
     validate_user_and_password
@@ -41,7 +44,7 @@ action :create do
     options['MO'] = @new_resource.frequency_modifier if frequency_modifier_allowed.include?(@new_resource.frequency)
     options['SD'] = @new_resource.start_day unless @new_resource.start_day.nil?
     options['ST'] = @new_resource.start_time unless @new_resource.start_time.nil?
-    options['TR'] = "\"#{@new_resource.command}\" "
+    options['TR'] = @new_resource.command
     options['RU'] = @new_resource.user
     options['RP'] = @new_resource.password if use_password?
     options['RL'] = 'HIGHEST' if @new_resource.run_level == :highest
@@ -50,6 +53,7 @@ action :create do
     options['M'] = @new_resource.months unless @new_resource.months.nil?
 
     run_schtasks 'CREATE', options
+    set_cwd(new_resource.cwd) if new_resource.cwd
     new_resource.updated_by_last_action true
     Chef::Log.info "#{@new_resource} task created"
   end
@@ -75,7 +79,7 @@ action :change do
     validate_interactive_setting
 
     options = {}
-    options['TR'] = "\"#{@new_resource.command}\" " if @new_resource.command
+    options['TR'] = @new_resource.command if @new_resource.command
     options['RU'] = @new_resource.user if @new_resource.user
     options['RP'] = @new_resource.password if @new_resource.password
     options['SD'] = @new_resource.start_day unless @new_resource.start_day.nil?
@@ -83,6 +87,7 @@ action :change do
     options['IT'] = '' if @new_resource.interactive_enabled
 
     run_schtasks 'CHANGE', options
+    set_cwd(new_resource.cwd) if new_resource.cwd != @current_resource.cwd
     new_resource.updated_by_last_action true
     Chef::Log.info "Change #{@new_resource} task ran"
   else
@@ -112,7 +117,7 @@ action :end do
     end
   else
     Chef::Log.fatal "#{@new_resource} task doesn't exist - nothing to do"
-    fail Errno::ENOENT, "#{@new_resource}: task does not exist, cannot end"
+    raise Errno::ENOENT, "#{@new_resource}: task does not exist, cannot end"
   end
 end
 
@@ -127,7 +132,7 @@ action :enable do
     end
   else
     Chef::Log.fatal "#{@new_resource} task doesn't exist - nothing to do"
-    fail Errno::ENOENT, "#{@new_resource}: task does not exist, cannot enable"
+    raise Errno::ENOENT, "#{@new_resource}: task does not exist, cannot enable"
   end
 end
 
@@ -147,20 +152,21 @@ end
 
 def load_current_resource
   @current_resource = Chef::Resource::WindowsTask.new(@new_resource.name)
-  @current_resource.task_name(@new_resource.task_name)
+  pathed_task_name = @new_resource.task_name.start_with?('\\') ? @new_resource.task_name : "\\#{@new_resource.task_name}"
 
-  pathed_task_name = @new_resource.task_name[0, 1] == '\\' ? @new_resource.task_name : @new_resource.task_name.prepend('\\')
-  task_hash = load_task_hash(@current_resource.task_name)
-  if task_hash[:TaskName] == pathed_task_name
+  @current_resource.task_name(pathed_task_name)
+  task_hash = load_task_hash(pathed_task_name)
+
+  if task_hash.respond_to?(:[]) && task_hash[:TaskName] == pathed_task_name
     @current_resource.exists = true
     @current_resource.status = :running if task_hash[:Status] == 'Running'
     if task_hash[:ScheduledTaskState] == 'Enabled'
       @current_resource.enabled = true
     end
-    @current_resource.cwd(task_hash[:Folder])
+    @current_resource.cwd(task_hash[:StartIn]) unless task_hash[:StartIn] == 'N/A'
     @current_resource.command(task_hash[:TaskToRun])
     @current_resource.user(task_hash[:RunAsUser])
-  end if task_hash.respond_to? :[]
+  end
 end
 
 private
@@ -168,7 +174,8 @@ private
 def run_schtasks(task_action, options = {})
   cmd = "schtasks /#{task_action} /TN \"#{@new_resource.task_name}\" "
   options.keys.each do |option|
-    cmd += "/#{option} #{options[option]} "
+    cmd += "/#{option} "
+    cmd += "\"#{options[option]}\" " unless options[option] == ''
   end
   Chef::Log.debug('running: ')
   Chef::Log.debug("    #{cmd}")
@@ -177,12 +184,52 @@ end
 
 def task_need_update?
   # gsub needed as schtasks converts single quotes to double quotes on creation
-  @current_resource.command != @new_resource.command.tr("'", "\"") ||
+  @current_resource.command != @new_resource.command.tr("'", '"') ||
     @current_resource.user != @new_resource.user
 end
 
-def load_task_hash(task_name)
+def set_cwd(folder)
   Chef::Log.debug 'looking for existing tasks'
+
+  # we use shell_out here instead of shell_out! because a failure implies that the task does not exist
+  xml_cmd = shell_out("schtasks /Query /TN \"#{@new_resource.task_name}\" /XML")
+
+  return if xml_cmd.exitstatus != 0
+
+  doc = REXML::Document.new(xml_cmd.stdout)
+
+  Chef::Log.debug 'Removing former CWD if any'
+  doc.root.elements.delete('Actions/Exec/WorkingDirectory')
+
+  unless folder.nil?
+    Chef::Log.debug 'Setting CWD as #folder'
+    cwd_element = REXML::Element.new('WorkingDirectory')
+    cwd_element.add_text(folder)
+    exec_element = doc.root.elements['Actions/Exec']
+    exec_element.add_element(cwd_element)
+  end
+
+  temp_task_file = ::File.join(ENV['TEMP'], 'windows_task.xml')
+  begin
+    ::File.open(temp_task_file, 'w:UTF-16LE') do |f|
+      doc.write(f)
+    end
+
+    options = {}
+    options['RU'] = @new_resource.user if @new_resource.user
+    options['RP'] = @new_resource.password if @new_resource.password
+    options['IT'] = '' if @new_resource.interactive_enabled
+    options['XML'] = temp_task_file
+
+    run_schtasks('DELETE', 'F' => '')
+    run_schtasks('CREATE', options)
+  ensure
+    ::File.delete(temp_task_file)
+  end
+end
+
+def load_task_hash(task_name)
+  Chef::Log.debug 'Looking for existing tasks'
 
   # we use shell_out here instead of shell_out! because a failure implies that the task does not exist
   output = shell_out("schtasks /Query /FO LIST /V /TN \"#{task_name}\"").stdout
@@ -203,7 +250,7 @@ def load_task_hash(task_name)
   task
 end
 
-SYSTEM_USERS = ['NT AUTHORITY\SYSTEM', 'SYSTEM', 'NT AUTHORITY\LOCALSERVICE', 'NT AUTHORITY\NETWORKSERVICE']
+SYSTEM_USERS = ['NT AUTHORITY\SYSTEM', 'SYSTEM', 'NT AUTHORITY\LOCALSERVICE', 'NT AUTHORITY\NETWORKSERVICE'].freeze
 
 def validate_user_and_password
   if @new_resource.user && use_password?
@@ -222,13 +269,13 @@ end
 def validate_create_day
   return unless @new_resource.day
   unless [:weekly, :monthly].include?(@new_resource.frequency)
-    fail 'day attribute is only valid for tasks that run weekly or monthly'
+    raise 'day attribute is only valid for tasks that run weekly or monthly'
   end
   if @new_resource.day.is_a? String
     days = @new_resource.day.split(',')
     days.each do |day|
       unless ['mon', 'tue', 'wed', 'thu', 'fri', 'sat', 'sun', '*'].include?(day.strip.downcase)
-        fail 'day attribute invalid.  Only valid values are: MON, TUE, WED, THU, FRI, SAT, SUN and *.  Multiple values must be separated by a comma.'
+        raise 'day attribute invalid.  Only valid values are: MON, TUE, WED, THU, FRI, SAT, SUN and *.  Multiple values must be separated by a comma.'
       end
     end
   end
@@ -237,13 +284,13 @@ end
 def validate_create_months
   return unless @new_resource.months
   unless [:monthly].include?(@new_resource.frequency)
-    fail 'months attribute is only valid for tasks that run monthly'
+    raise 'months attribute is only valid for tasks that run monthly'
   end
   if @new_resource.months.is_a? String
     months = @new_resource.months.split(',')
     months.each do |month|
       unless ['JAN', 'FEB', 'MAR', 'APR', 'MAY', 'JUN', 'JUL', 'AUG', 'SEP', 'OCT', 'NOV', 'DEC', '*'].include?(month.strip.upcase)
-        fail 'months attribute invalid. Only valid values are: JAN, FEB, MAR, APR, MAY, JUN, JUL, AUG, SEP, OCT, NOV, DEC and *. Multiple values must be separated by a comma.'
+        raise 'months attribute invalid. Only valid values are: JAN, FEB, MAR, APR, MAY, JUN, JUL, AUG, SEP, OCT, NOV, DEC and *. Multiple values must be separated by a comma.'
       end
     end
   end
@@ -256,23 +303,23 @@ def validate_create_frequency_modifier
     case @new_resource.frequency
     when :minute
       unless @new_resource.frequency_modifier.to_i > 0 && @new_resource.frequency_modifier.to_i <= 1439
-        fail "frequency_modifier value #{@new_resource.frequency_modifier} is invalid.  Valid values for :minute frequency are 1 - 1439."
+        raise "frequency_modifier value #{@new_resource.frequency_modifier} is invalid.  Valid values for :minute frequency are 1 - 1439."
       end
     when :hourly
       unless @new_resource.frequency_modifier.to_i > 0 && @new_resource.frequency_modifier.to_i <= 23
-        fail "frequency_modifier value #{@new_resource.frequency_modifier} is invalid.  Valid values for :hourly frequency are 1 - 23."
+        raise "frequency_modifier value #{@new_resource.frequency_modifier} is invalid.  Valid values for :hourly frequency are 1 - 23."
       end
     when :daily
       unless @new_resource.frequency_modifier.to_i > 0 && @new_resource.frequency_modifier.to_i <= 365
-        fail "frequency_modifier value #{@new_resource.frequency_modifier} is invalid.  Valid values for :daily frequency are 1 - 365."
+        raise "frequency_modifier value #{@new_resource.frequency_modifier} is invalid.  Valid values for :daily frequency are 1 - 365."
       end
     when :weekly
       unless @new_resource.frequency_modifier.to_i > 0 && @new_resource.frequency_modifier.to_i <= 52
-        fail "frequency_modifier value #{@new_resource.frequency_modifier} is invalid.  Valid values for :weekly frequency are 1 - 52."
+        raise "frequency_modifier value #{@new_resource.frequency_modifier} is invalid.  Valid values for :weekly frequency are 1 - 52."
       end
     when :monthly
       unless ('1'..'12').to_a.push('FIRST', 'SECOND', 'THIRD', 'FOURTH', 'LAST', 'LASTDAY').include?(@new_resource.frequency_modifier.to_s.upcase)
-        fail "frequency_modifier value #{@new_resource.frequency_modifier} is invalid.  Valid values for :monthly frequency are 1 - 12, 'FIRST', 'SECOND', 'THIRD', 'FOURTH', 'LAST', 'LASTDAY'."
+        raise "frequency_modifier value #{@new_resource.frequency_modifier} is invalid.  Valid values for :monthly frequency are 1 - 12, 'FIRST', 'SECOND', 'THIRD', 'FOURTH', 'LAST', 'LASTDAY'."
       end
     end
   end