+++ test/cookbooks/minimal/recipes/feature.rb
@@ -1,0 +1,5 @@
+node.default['windows']['feature_provider'] = 'dism'
+
+windows_feature 'TelnetClient' do
+  action :install
+end