+++ test/cookbooks/minimal/recipes/tasks.rb
@@ -47,8 +47,12 @@ end
 
 windows_task 'create task delete_me' do
   name 'delete_me'
+  user 'NT AUTHORITY\SYSTEM'
+  run_level :highest
+  password 'ignored'
   action :create
   command 'dir'
+  notifies :create, 'file[c:/notifytest.txt]', :immediately
 end
 
 windows_task 'delete task delete_me' do
@@ -58,6 +62,13 @@ end
 
 windows_task 'task_for_system' do
   command 'dir'
-  user 'system'
-  password 'ignored'
+  run_level :highest
+  user 'vagrant'
+  password 'vagrant'
+  cwd ENV['TEMP']
+end
+
+file 'c:/notifytest.txt' do
+  content 'blah'
+  action :nothing
 end