+++ test/integration/feature/pester/minimal.package.Tests.ps1
@@ -1,0 +1,10 @@
+$global:progressPreference = 'SilentlyContinue'
+
+describe 'minimal::feature' {
+  context 'minimal_feature' {
+
+    it "feature TelnetClient was created"  {
+      Get-Command Telnet -ErrorAction SilentlyContinue | Should Not Be $Null
+    }
+  }
+}